refactor: playground in feature oriented architecture
