+++ src/lib/features/playground/feature-evaluator/index.ts
@@ -5,6 +5,6 @@ import { ClientFeaturesResponse } from './feature';
 import InMemStorageProvider from './repository/storage-provider-in-mem';
 
 // exports
-export { Strategy } from './strategy/index';
+export { Strategy } from './strategy';
 export { Context, Variant, FeatureEvaluator, InMemStorageProvider };
 export type { ClientFeaturesResponse, FeatureEvaluatorConfig };
